pkg/components/util: split tests between _test and internal
To promote only testing exported interfaces. Unfortunately
TestChartFromManifests cannot be tested with exported interface, as
InstallComponent using it must have valid kubeconfig to work with.
